refactor: move TableContainers to dedicated folder
