+++ pages/payments/index.tsx
@@ -3,7 +3,7 @@ import supertokensNode from 'supertokens-node'
 import * as SuperTokensConfig from '../../config/backendConfig'
 import Session from 'supertokens-node/recipe/session'
 import { GetServerSideProps } from 'next'
-import TableContainer from '../../components/TableContainer'
+import TableContainer from '../../components/TableContainer/TableContainer'
 import { ButtonDisplayData } from 'redis/dashboardCache'
 import Image from 'next/image'
 import Link from 'next/link'
